requirements.txt
flask
openai
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
json
re
